requirements.txt
aiofile==3.7.2
aiofiles==0.7.0
aiosqlite==0.17.0
anyio==3.3.4
aredis==1.1.8
argon2-cffi==21.1.0
asgiref==3.4.1
async-generator==1.10
asyncpg==0.24.0
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
black==21.10b0
caio==0.9.3
captcha==0.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
databases==0.5.3
decorator==5.1.0
environs==9.3.4
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.1
httptools==0.3.0
httpx==0.21.0
idna==3.3
iniconfig==1.1.1
iso8601==0.1.16
jedi==0.18.0
marshmallow==3.14.0
matplotlib-inline==0.1.3
multidict==5.2.0
mypy-extensions==0.4.3
packaging==21.2
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.5
pytest-sanic==1.9.1
python-dotenv==0.19.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
rfc3986==1.5.0
sanic==21.9.1
sanic-envconfig==1.3.4
sanic-jwt==1.7.0
sanic-routing==0.7.2
secure==0.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
toml==0.10.2
tomli==1.2.2
tortoise-orm==0.17.8
traitlets==5.1.1
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
v==0.0.0
wcwidth==0.2.5
websockets==10.0
